backend/requirements.txt
fastapi>=0.115.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
uvicorn[standard]>=0.30.0,<1.0.0
